+++ src/main/java/by/bsu/rfict/WarehouseAccountingSystem/entity/Item.java
@@ -26,7 +26,7 @@ public class Item {
     private int cost;
 
     @Column(name = "weight")
-    private short weight;
+    private double weight;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id", referencedColumnName = "id")
